requirements.txt
fritzconnection
paho-mqtt
pyyaml